+++ src/main/java/com/example/demo/repositories/PersonnelRepository.java
@@ -4,5 +4,9 @@ import com.example.demo.entities.Personnel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PersonnelRepository extends JpaRepository<Personnel, Long> {
-    Personnel findByUserId(Integer userId);
+    Personnel findByUserId(Long userId);
+
+
+
+
 }
